### 2023-12-04 #6114 @editorialbot
[PRE REVIEW]: Systems Neuro Browser (SNUB)
**Submitting author:** @calebweinreb (<a href="http://orcid.org/0000-0001-6100-6084">Caleb Weinreb</a>)
**Repository:** https://github.com/calebweinreb/SNUB
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.3
