### 2023-12-05 #6114 @emdupre
Thank you, @calebweinreb !

While I look through these suggestions, could you please also address this EditorialBot comment ?

> Failed to discover a Statement of need section in paper

You can find more information on what we expect in JOSS submissions [in our documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) on "What should my paper contain?".
