+++ paper.md
@@ -20,29 +20,21 @@ bibliography: paper.bib
 
 ---
 
+
 # Summary
 
-SNUB is a tool for exploring time-series data, such as neural 
-recordings, behavioral videos, temperature, movement or other sensor signals, 
-and higher-level annotations derived from such data. The interface is 
-divided into synchronized windows that each show a different data stream. 
-The linked data views allow users to quickly inspect the relationships between
-experimental phenomena, such as the behaviors that occur during a particular 
-pattern of neural activity (\autoref{fig:screenshot}). 
+A core goal of neuroscience is to discover temporal patterns in behavior and neurophysiology. Though a variety of tools exist to characterize these relationships, there is still no substitute for direct inspection as a way to notice unexpected patterns and generate new hypotheses. To facilitate this process, we have developed the Systems Neuro Browser (SNUB), a graphical user interface for exploring time-series data. SNUB is a flexible, general-purpose tool that allows users to build a dashboard of synchronized data views, including neural recordings, behavioral videos, and annotations derived from these data. 
+
+
+# Statement of need
+
+Direct inspection of behavior and neurophysiology recordings is hard because the data are typically high-dimensional, come in a variety of modalities (such as raw video, pose tracking, spike trains, calcium traces, etc.) with different sampling rates and methods of visualization. SNUB lowers the activation energy for data exploration by integrating these data streams into a single easy-to-navigate interface. The interface is divided into synchronized windows that each show a different data stream. The linked data views allow users to quickly inspect the relationships between experimental phenomena, such as the behaviors that occur during a particular pattern of neural activity (\autoref{fig:screenshot}). 
 
 ![Screenshot from SNUB.\label{fig:screenshot}](docs/media/screenshot.png)
 
-We provide dedicated widgets and loading functions for exploring
-raw video, 3D animal pose, behavior annotations, electrophysiology recordings,
-and calcium imaging data - either as a raster or as a super-position of 
-labeled regions of interest (ROIs). More broadly, SNUB can dislay any data
-that takes the form of a heatmap, scatter plot, video, or collection of 
-named temporally-varying signals. 
+We provide dedicated widgets and loading functions for exploring raw video, 3D animal pose, behavior annotations, electrophysiology recordings, and calcium imaging data - either as a raster or as a super-position of labeled regions of interest (ROIs). More broadly, SNUB can dislay any data that takes the form of a heatmap, scatter plot, video, or collection of named temporally-varying signals. 
 
-In addition to the front-end GUI, we include a library of functions for
-ingesting raw data and saving it to a format that is readable by the SNUB
-viewer. The following code, for example, creates a project with paired 
-electrophysiology and video data.
+In addition to the front-end GUI, we include a library of functions for ingesting raw data and saving it to a format that is readable by the SNUB viewer. The following code, for example, creates a project with paired electrophysiology and video data.
 
 ```
 snub.io.create_project(project_directory, duration=1800)
@@ -50,26 +42,12 @@ snub.io.add_video(project_directory, 'path/to/my_video.avi', name='IR_camera')
 snub.io.add_splikeplot(project_directory, 'my_ephys_data', spike_times, spike_labels)
 ```
 
-SNUB is a flexible general-purpose tool that complements more specialized 
-packages such as rastermap [@rastermap] and Bento [@bento]. The rastermap
-interface, for example, is hard-coded for the display of neural activity 
-rasters, ROIs and 2D embeddings of neural activity. Bento is hard-coded for 
-the display of neural activity rasters, behavioral videos and behavioral 
-annotations. SNUB can reproduce either of these configurations and is 
[…]
