### 2023-12-13 #6114 @emdupre
👋 Hi @vigji, @niksirbi, and @tgeijten,

Would you be willing to review _Systems Neuro Browser (SNUB)_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/calebweinreb/SNUB) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.06114/joss.06114/10.21105.joss.06114.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer.

🌟 Generally, we ask that all reviews are completed within six weeks, although I know that many of us are out-of-office during the holiday season. So please let me know if you'd be available on this six week timeline, or if you'd need a slightly longer window.

Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you for considering !
