### 2023-12-14 #6114 @niksirbi
👋 Hi @emdupre, thanks for pinging me.

SNUB looks very interesting and relevant to my work, I'll be happy to review it!
Six weeks sounds feasible, so sign me up.

Thanks,
Niko
