### 2024-01-04 calebweinreb/SNUB#5 @vigji
Opening review
I am starting here review for JOSS, will make several issues to have separate conversations on different improvement suggestions so that we could bounce things back and forth before I report the review outcome on the JOSS issue.
