### 2024-01-04 calebweinreb/SNUB#6 @vigji
MacOS installation problems
I tried to install the package on a macOS M1 computer in a fresh conda environment using python `3.8`, `3.9`, `3.10` and for all of them I keep getting

```shell
bus error
```

Every time I try to either run snub from the terminal or import anything from snub in a script.

This might be a OS issue on my side although it does suggest there might be compatibility issues. Which operative systems has it been tested on? Is this an issue already experienced by any user?

I will try to work out some solution and report back if I find one. I will test on a Windows machine in some days.
