### 2024-01-04 calebweinreb/SNUB#5 @vigji
As I have some installation issues (see #6) I will start from points that are not strictly related to code usage yet!
