### 2024-01-04 calebweinreb/SNUB#7 @vigji
[Minor] heavy codebase
This seems to have been raised already (#3) but the issue is still present even if the old issue has been closed as completed
