### 2024-01-04 calebweinreb/SNUB#8 @vigji
[Major] No testing
The repo currently has no test modules, which is a major flow and would be highly recommended for deploying reliable and maintainable code.

Would you consider add testing modules to the codebase, and integrate a testing workflow using GitHub Actions?
