### 2024-01-04 calebweinreb/SNUB#9 @vigji
[Major] No support for standard neurodata formats
As the package aims at helping out people dealing with their own recordings and experiments, it would greatly benefit from supporting standard neuro data formats.

I believe that the impact of the tool would be way larger if there were some more integration with the existing ecosystem of data formats, in the current shape it feels a bit a lab-specific tool (format-wise), instead of forcing the user to structure their dataset to the highly specific requirements of the package. I could recommend for example look for at least proof-of-principle integration with NWB, or SpikeInterface.

A good way of smoothing this up would be to showcase the software on openly accessible datasets, instead of the custom datasets provided as examples.
