### 2024-01-04 calebweinreb/SNUB#10 @vigji
[Minor] DOI for datasets
Currently, documentation points at dropbox folders to download the example datasets. Maybe I am wrong, but I believe that does links do not have the stability that would be warranted by a DOI-minting data storage.

I would strongly advise to keep the data on places like Zenodo or Figshare and link them via DOI addresses.
