### 2024-01-04 calebweinreb/SNUB#11 @vigji
[Minor] conflation of data and GUI parameters logic
I have been reading the functions for the project creation, and I am bit confused by the conflation of the (thin) data-munging logic and the visualization settings logic in the same functions.

Maybe it would be better to separate the part where one links to the project all the available data and the part where one specifies the default parameters for the GUI.
