### 2024-01-04 calebweinreb/SNUB#12 @vigji
[Minor] data copy logic
Since there is such a minimal data manipulation (mostly, just rasterisation of the spike times and traces sorting, if I am correct), I do not understand the need to copy all the files in a project data folder.
I can see that some data types offer the option of not been copied, just linked, but not all of them. But this is not only an issue for space, also for processing reasons. For example, it would force the user into having to re-run project generation if they change anything in the data preprocessing (eg, spike-detection or calcium signal extraction) which can be annoying and a source of mistakes.
