### 2024-01-04 calebweinreb/SNUB#13 @vigji
[Major] package deployment via pip
Since there are no difficult dependencies in the project (and no environment file necessary as a result), I would highly recommend distribute the package via PyPI/`pip`, so that it can be more easily installed and kept updated by all users.
