### 2024-01-04 calebweinreb/SNUB#14 @vigji
[Minor] Icon attribution
As far as I can see, the icon is a schematic from scidraw (full disclosure, one I draw!), and [the licence](https://creativecommons.org/licenses/by/4.0/#ref-appropriate-credit) under which the image is free to use would require a small attribution somewhere.
