### 2024-01-04 calebweinreb/SNUB#15 @vigji
[Minor] Linting
The code seems to not have been passed through any linting tool (such as flake8) to remove minor departures from PEP recommendations (eg, unused imports, unused arguments in functions). To make it publishable-quality, I would recommend including some linting tool in the workflow to iron those out.
