### 2024-01-05 #6114 @niksirbi
Happy New Year!

Confirming that I'm happy to start reviewing SNUB next week.
