### 2024-01-06 #6114 @vigji
Happy new year! I've start opening issues on SNUB repo to discuss with the author, I'll link everything here in a final review at the end.
