### 2024-01-06 calebweinreb/SNUB#12 @calebweinreb
This is a good point and closely related to another issue you raised (https://github.com/calebweinreb/SNUB/issues/11) about conflation of data and GUI parameters. The rationale for both of these design choices is that SNUB is almost entirely focused on visualization (as opposed to data processing/transformation) and designed to be as fast/responsive as possible. It is meant to be somewhat analogous to a plotting library, e.g. with `plt.scatter` replaced by `snub.io.add_scatter` (though a few data processing utils such as wrappers for UMAP or rastermap are included just for convenience. The rationale for re-saving data in a fixed, deterministic format is that it allows SNUB to be agnostic about the original format that the data comes in, and to avoid performing costly conversion and/or preprocessing-steps every time a dataset is loaded.

The primary use case we have in mind is that a person has collected data and begun to analyze it in jupyter. They are totally comfortable loading the data and manipulating it with python. They likely are already generating static versions of SNUB-like data views, such as aligned heatmaps and trace plots, and now would like a frictionless way to pan/zoom around these plots and keep all the data views linked together. So instead of a big matplotlib figure, they create a SNUB project. Just as the matplotlib figure is agnostic to any loading/preprocessing that has already happened, so are the SNUB i/o routines.

Let me know if this clarifies both issues, and if you think it could be explained better in the paper. And thank you for the careful review!
