### 2024-01-06 calebweinreb/SNUB#7 @calebweinreb
This is a great point - the earlier issue was closed without addressing the problem. I deleted the docs build and some unused gifs/videos from the repo history and brought the total size down to ~30Mb (from ~400Mb)
