+++ snub/gui/tracks/spike.py
@@ -1,46 +1,69 @@
 from PyQt5.QtCore import *
 from PyQt5.QtWidgets import *
 from PyQt5.QtGui import *
-from functools import partial
 import os
 import numpy as np
 import cmapy
-import time
 
 from vispy.scene import SceneCanvas
 from vispy.scene.visuals import Markers, Line
-from snub.gui.tracks import Track, TracePlot, TrackGroup, Heatmap
+from snub.gui.tracks import TracePlot, TrackGroup, Heatmap
 
 
-'''
+"""
 class SpikePlot(Heatmap):
     def __init__(self, config, selected_intervals, spikes_path=None, heatmap_path=None, **kwargs):
         print(spikes_path, heatmap_path)
         super().__init__(config, selected_intervals, data_path=heatmap_path, **kwargs)
         self.spike_data = np.load(os.path.join(config['project_directory'],spikes_path))
-'''
+"""
 
 
 class SpikePlot(Heatmap):
-    def __init__(self, config, selected_intervals, spikes_path=None, markersize=5,
-                       heatmap_path=None, heatmap_range=60, colormap='viridis', **kwargs):
-
+    def __init__(
+        self,
+        config,
+        selected_intervals,
+        spikes_path=None,
+        markersize=5,
+        heatmap_path=None,
+        heatmap_range=60,
+        colormap="viridis",
+        **kwargs
+    ):
         super().__init__(config, selected_intervals, data_path=heatmap_path, **kwargs)
         self.heatmap_range = heatmap_range
-        spike_data = np.load(os.path.join(config['project_directory'],spikes_path))
-        self.spike_times,self.spike_labels = spike_data[:,0], spike_data[:,1].astype(int)
+        spike_data = np.load(os.path.join(config["project_directory"], spikes_path))
+        self.spike_times, self.spike_labels = spike_data[:, 0], spike_data[:, 1].astype(
+            int
+        )
         self.max_label = self.spike_labels.max()
-        self.markersize=markersize
+        self.markersize = markersize
         self.colormap = colormap
-        self.cmap = cmapy.cmap(self.colormap).squeeze()[:,::-1].astype(np.float32)/255
-        self.canvas = SceneCanvas(self, keys='interactive', bgcolor=self.cmap[0], show=True)
-        self.viewbox = self.canvas.central_widget.add_grid().add_view(row=0, col=0, camera='panzoom')
-        line_verts = np.vstack([
-            np.ones(self.max_label)*self.spike_times.min()-10,
-            np.arange(self.max_label),
-            np.ones(self.max_label)*self.spike_times.max()+10,
-            np.arange(self.max_label)]).T
-        self.lines = Line(pos=line_verts.reshape(-1,2), color=np.clip(self.cmap[0]+.1,0,1), method='gl', width=0.5, connect='segments')
+        self.cmap = (
+            cmapy.cmap(self.colormap).squeeze()[:, ::-1].astype(np.float32) / 255
+        )
+        self.canvas = SceneCanvas(
+            self, keys="interactive", bgcolor=self.cmap[0], show=True
+        )
+        self.viewbox = self.canvas.central_widget.add_grid().add_view(
+            row=0, col=0, camera="panzoom"
+        )
+        line_verts = np.vstack(
+            [
+                np.ones(self.max_label) * self.spike_times.min() - 10,
+                np.arange(self.max_label),
+                np.ones(self.max_label) * self.spike_times.max() + 10,
+                np.arange(self.max_label),
+            ]
+        ).T
+        self.lines = Line(
+            pos=line_verts.reshape(-1, 2),
+            color=np.clip(self.cmap[0] + 0.1, 0, 1),
+            method="gl",
+            width=0.5,
[…]
