### 2024-01-06 calebweinreb/SNUB#15 @calebweinreb
Thanks for the suggestion! I used black to reformat the code and VS code to highlight unused imports (which were deleted) (https://github.com/calebweinreb/SNUB/commit/aee2d2adb4a5eba6938550e89898c60b1dac2162). I know some projects also include these kinds of checks as github actions but I think that might be overkill here as I am currently the only active developer.
