+++ setup.cfg
@@ -1,0 +1,43 @@
+[metadata]
+name = snub
+author = Caleb Weinreb
+author_email = calebsw@gmail.com
+url = https://github.com/dattalab/SNUB
+classifiers =
+    Programming Language :: Python :: 3
+    Operating System :: OS Independent
+
+[options]
+entry_points = {"console_scripts": ["snub = snub.gui.main:run"]}
+packages = find:
+include_package_data = True
+python_requires = >=3.8
+install_requires =
+    PyQt5
+    numpy
+    scikit-learn
+    tqdm
+    cmapy
+    interlap
+    numba
+    vispy
+    imageio
+    imageio-ffmpeg
+    umap-learn
+    rastermap
+    ipykernel
+    pyqtgraph
+    networkx
+    opencv-python-headless
+    vidio>=0.0.3
+
+[options.package_data]
+* = *.md
+
+[versioneer]
+VCS = git
+style = pep440
+versionfile_source = snub/_version.py
+versionfile_build = snub/_version.py
+tag_prefix =
+parentdir_prefix = 
