### 2024-01-06 7db6132
Proper formatting of entrypoints in setup.cfg
