+++ setup.cfg
@@ -8,7 +8,6 @@ classifiers =
     Operating System :: OS Independent
 
 [options]
-entry_points = {"console_scripts": ["snub = snub.gui.main:run"]}
 packages = find:
 include_package_data = True
 python_requires = >=3.8
@@ -30,6 +29,10 @@ install_requires =
     networkx
     opencv-python-headless
     vidio>=0.0.3
+
+[options.entry_points]
+console_scripts =
+    snub = snub.gui.main:run
 
 [options.package_data]
 * = *.md
