### 2024-01-06 167511c
Propert attribution for app icon
