### 2024-01-06 calebweinreb/SNUB#14 @calebweinreb
My apologies for the oversight, and thanks for contributing the drawing! I added an acknowledgement in the paper (https://github.com/calebweinreb/SNUB/commit/167511c5bbe5b6c701d718c2b7cd85d9f6915d7a).
