### 2024-01-08 #6114 @emdupre
Thank you for confirming, @vigji, @niksirbi, and @lucasmiranda42 !

@lucasmiranda42, I'll add you as a reviewer now and then open the deviated review issue !

We now have a sufficient number of reviewers for _SNUB_, but thank you for considering this invitation, @jbohnslav 🌻
