### 2024-01-08 calebweinreb/SNUB#6 @calebweinreb
Hello, thanks for raising this!

I tried installing on an M1 and M2 mac without any issues. Was there any other terminal output? Or could SNUB be interfering with another program's memory access?
