### 2024-01-08 calebweinreb/SNUB#6 @vigji
Uhm, intriguing! This is the only terminal output I get when it fails, no matter where I try (terminal / VS code), if I run `snub` from the environment where I installed the package:
```
[1]    28246 bus error  snub
```

This is the full listing of `pip list` if it can be informative:
```
Package                Version  Editable project location
---------------------- -------- -------------------------
appnope                0.1.3
asttokens              2.4.1
backcall               0.2.0
cmapy                  0.6.6
comm                   0.2.1
contourpy              1.1.1
cycler                 0.12.1
debugpy                1.8.0
decorator              5.1.1
executing              2.0.1
fonttools              4.47.0
freetype-py            2.4.0
h5py                   3.10.0
hsluv                  5.0.4
imageio                2.33.1
imageio-ffmpeg         0.4.9
importlib-metadata     7.0.1
importlib-resources    6.1.1
interlap               0.2.7
ipykernel              6.28.0
ipython                8.12.3
jedi                   0.19.1
joblib                 1.3.2
jupyter_client         8.6.0
jupyter_core           5.7.0
kiwisolver             1.4.5
llvmlite               0.41.1
matplotlib             3.7.4
matplotlib-inline      0.1.6
natsort                8.4.0
nest-asyncio           1.5.8
networkx               3.1
numba                  0.58.1
numpy                  1.24.4
opencv-python          4.9.0.80
opencv-python-headless 4.9.0.80
packaging              23.2
parso                  0.8.3
pexpect                4.9.0
pickleshare            0.7.5
pillow                 10.2.0
pip                    23.3.2
platformdirs           4.1.0
prompt-toolkit         3.0.43
psutil                 5.9.7
ptyprocess             0.7.0
pure-eval              0.2.2
Pygments               2.17.2
pynndescent            0.5.11
pyparsing              3.1.1
PyQt5                  5.15.10
PyQt5-Qt5              5.15.12
PyQt5-sip              12.13.0
pyqtgraph              0.13.3
python-dateutil        2.8.2
pyzmq                  25.1.2
rastermap              0.9.4
scikit-learn           1.3.2
scipy                  1.10.1
setuptools             69.0.3
six                    1.16.0
snub                   0.0.3    /Users/vigji/code/SNUB
stack-data             0.6.3
threadpoolctl          3.2.0
tornado                6.4
tqdm                   4.66.1
traitlets              5.14.1
typing_extensions      4.9.0
umap-learn             0.5.5
vidio                  0.0.4
vispy                  0.14.1
wcwidth                0.2.12
wheel                  0.42.0
zipp                   3.17.0
```

Can you export a `environment.yml` file so that I can recreate your environment as close as possible?
