### 2024-01-09 calebweinreb/SNUB#6 @niksirbi
For what it's worth, I've followed the installation instructions and they worked as advertised on my M2 Mac (in a conda environment with Python 3.8).
