### 2024-01-09 calebweinreb/SNUB#8 @niksirbi
Agreed, this is a major issue.
I recommend at least starting with some smoke tests for the most important functionalities/workflows.
