### 2024-01-09 calebweinreb/SNUB#13 @niksirbi
Agreed, deployment on pip would be feasible and could enable a one-line install (`pip install snub`) on most platforms. It's worth pursuing sooner rather than later since it will make SNUB much more accessible.

Once the first release is deployed, update releases can be automated via GitHub Actions. Some example action workflows can be found [here](https://github.com/neuroinformatics-unit/actions).
