### 2024-01-10 calebweinreb/SNUB#6 @calebweinreb
Here's a conda env config for an M1 where it worked
[requirements.txt](https://github.com/calebweinreb/SNUB/files/13888880/requirements.txt)
