### 2024-01-17 5d2e1fa
added basic test suite
