+++ setup.cfg
@@ -23,7 +23,7 @@ install_requires =
     imageio
     imageio-ffmpeg
     umap-learn
-    rastermap
+    rastermap==0.1.3
     ipykernel
     pyqtgraph
     networkx
