### 2024-01-17 cd8b678
add deps for testing
