+++ setup.cfg
@@ -29,6 +29,8 @@ install_requires =
     networkx
     opencv-python-headless
     vidio>=0.0.3
+    pytest
+    pytest-qt
 
 [options.entry_points]
 console_scripts =
