### 2024-01-19 #6187 @emdupre
👋  Hi @vigji, I noticed that you hadn't yet created your reviewer checklist, so I wanted to make sure you hadn't hit any issues in doing so. If there's anything I can clarify, please let me know !

I also wanted to cross-link the issues you've already opened on _SNUB_ (thank you !) :

- https://github.com/calebweinreb/SNUB/issues/5
- https://github.com/calebweinreb/SNUB/issues/6
- https://github.com/calebweinreb/SNUB/issues/7
- https://github.com/calebweinreb/SNUB/issues/8
- https://github.com/calebweinreb/SNUB/issues/9
- https://github.com/calebweinreb/SNUB/issues/10
- https://github.com/calebweinreb/SNUB/issues/11
- https://github.com/calebweinreb/SNUB/issues/12
- https://github.com/calebweinreb/SNUB/issues/13
- https://github.com/calebweinreb/SNUB/issues/14
- https://github.com/calebweinreb/SNUB/issues/15

If&mdash;when you create the checklist&mdash;you want to directly link them with individual review points, that would help to structure the review ! But this isn't strictly necessary, as I can handle that post-hoc.
