### 2024-01-20 calebweinreb/SNUB#14 @vigji
No problem! Putting it in the paper is maybe to much as it conflates it with actual codebase sources, a very small link in the docs somewhere suffices if you prefer! But up to you :)
