### 2024-01-22 158ed52
added automatic plot generation from NWB (refactor of Mo's original approach)
