+++ snub/io/nwb/__init__.py
@@ -1,1 +1,0 @@
-from .convert import create_project_from_nwb, add_nwb
