+++ setup.cfg
@@ -31,6 +31,13 @@ install_requires =
     vidio>=0.0.3
     pytest
     pytest-qt
+    pynwb
+    ndx-pose
+    ndx-photometry
+    ndx-labels
+    ndx-depth-moseq
+    pandas
+
 
 [options.entry_points]
 console_scripts =

[…]
