### 2024-01-26 595c941
changed fixture scope to module
