+++ setup.cfg
@@ -37,6 +37,7 @@ install_requires =
     ndx-labels
     ndx-depth-moseq
     pandas
+    dandi
 
 
 [options.entry_points]
