### 2024-01-26 1f83ce6
updated NWB docs and added nwb test
