### 2024-01-28 41c1614
added NWB conversion tool description to the JOSS manuscript
