+++ paper.bib
@@ -1,102 +1,124 @@
 @software{vispy,
-  author       = {Luke Campagnola and
-                  Eric Larson and
-                  Almar Klein and
-                  David Hoese and
-                  Siddharth and
-                  Cyrille Rossant and
-                  Adam Griffiths and
-                  Nicolas P. Rougier and
-                  asnt and
-                  Kai Mühlbauer and
-                  Alexander Taylor and
-                  MSS and
-                  Talley Lambert and
-                  sylm21 and
-                  Alex J. Champandard and
-                  Max Hunter and
-                  Thomas Robitaille and
-                  Mustafa Furkan Kaptan and
-                  Elliott Sales de Andrade and
-                  Karl Czajkowski and
-                  Lorenzo Gaifas and
-                  Alessandro Bacchini and
-                  Guillaume Favelier and
-                  Etienne Combrisson and
-                  ThenTech and
-                  fschill and
-                  Mark Harfouche and
-                  Michael Aye and
-                  Casper van Elteren and
-                  Cedric GESTES},
-  title        = {vispy/vispy: Version 0.11.0},
-  month        = jul,
-  year         = 2022,
-  publisher    = {Zenodo},
-  version      = {v0.11.0},
-  doi          = {10.5281/zenodo.6795163},
-  url          = {https://doi.org/10.5281/zenodo.6795163}
+  author    = {Luke Campagnola and
+               Eric Larson and
+               Almar Klein and
+               David Hoese and
+               Siddharth and
+               Cyrille Rossant and
+               Adam Griffiths and
+               Nicolas P. Rougier and
+               asnt and
+               Kai Mühlbauer and
+               Alexander Taylor and
+               MSS and
+               Talley Lambert and
+               sylm21 and
+               Alex J. Champandard and
+               Max Hunter and
+               Thomas Robitaille and
+               Mustafa Furkan Kaptan and
+               Elliott Sales de Andrade and
+               Karl Czajkowski and
+               Lorenzo Gaifas and
+               Alessandro Bacchini and
+               Guillaume Favelier and
+               Etienne Combrisson and
+               ThenTech and
+               fschill and
+               Mark Harfouche and
+               Michael Aye and
+               Casper van Elteren and
+               Cedric GESTES},
+  title     = {vispy/vispy: Version 0.11.0},
+  month     = jul,
+  year      = 2022,
+  publisher = {Zenodo},
+  version   = {v0.11.0},
+  doi       = {10.5281/zenodo.6795163},
+  url       = {https://doi.org/10.5281/zenodo.6795163}
 }
 
 
 
 @article{bento,
[…]
