### 2024-01-28 calebweinreb/SNUB#9 @calebweinreb
Thanks for the suggestion!

Right now the software is relatively format-agnostic. Most of the io functions ingest numpy arrays directly, rather than paths to data in a specific format. So in a sense it is no more format-specific than matplotlib.

I agree that interfacing with standard formats would be useful through. To that end, we developed a conversion tool that generates SNUB datasets directly from NWB files (see https://github.com/calebweinreb/SNUB/pull/17). See documentation here: https://github.com/calebweinreb/SNUB/blob/dev/docs/source/nwb.rst
