### 2024-01-28 d428326
removed unused function from gui test module
