### 2024-01-28 calebweinreb/SNUB#8 @calebweinreb
Thanks for the suggestions! To address this, we have now added tests (https://github.com/calebweinreb/SNUB/pull/17). The test modules include:

- `test_gui.py`: Launches the GUI and loads a data-light example dataset with video, heatmap and trace plot. Currently this module does not include simulated GUI interactions.
- `test_io.py`: Creates a new SNUB project and runs through a basic project-building workflow, including addition of a heatmap, scatter plot and video. The test passes if these steps complete without error.
- `test_nwb.py`: Downloads an NWB file using the DANDI API and then converts the file into a SNUB dataset. The test passes if these steps complete without error. The test is essentially equivalent to the second example from the docs (https://github.com/calebweinreb/SNUB/blob/nwb/docs/source/nwb.rst)
