### 2024-01-28 bc06bb6
Added author, changed data on JOSS manuscript
