+++ paper.md
@@ -9,13 +9,16 @@ authors:
   - name: Caleb Weinreb
     orcid: 0000-0001-6100-6084
     affiliation: 1
+  - name: Mohammed Abdal Monium Osman
+    orcid: 0000-0001-8606-6518
+    affiliation: 1
   - name: Sandeep Robert Datta
     orcid: 0000-0002-8068-3862
     affiliation: 1
 affiliations:
  - name: Department of Neurobiology, Harvard University, Boston, MA
    index: 1
-date: 5 October 2022
+date: 28 January 2024
 bibliography: paper.bib
 
 ---
