### 2024-01-28 calebweinreb/SNUB#10 @calebweinreb
The example datasets have now been deposited on Zenodo https://doi.org/10.5281/zenodo.10578025
