### 2024-01-28 93e0569
changed name to systems-neuro-browser since this is unused on pypi
