+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = snub
+name = systems-neuro-browser
 author = Caleb Weinreb
 author_email = calebsw@gmail.com
 url = https://github.com/dattalab/SNUB
