### 2024-01-28 00ea2f1
remove duplicate name metadata from setup files
