+++ setup.cfg
@@ -33,7 +33,6 @@ install_requires =
     pynwb
     ndx-pose
     ndx-photometry
-    ndx-labels
     ndx-depth-moseq
     pandas
     dandi
