### 2024-01-29 e5114b5
added build info to readthedocs config
