### 2024-01-29 6389267
Updated install instructions to use PyPI
