### 2024-01-29 4f73a39
updated README to include install from pypi and NWB conversion, and link to readme in setup.cfg (for PyPI description)
