+++ setup.cfg
@@ -1,10 +1,15 @@
 [metadata]
 author = Caleb Weinreb
 author_email = calebsw@gmail.com
-url = https://github.com/dattalab/SNUB
+url = https://github.com/calebweinreb/SNUB
+description = Systems Neuro Browser
+long_description = file:README.md
+long_description_content_type = text/markdown
 classifiers =
     Programming Language :: Python :: 3
     Operating System :: OS Independent
+
+
 
 [options]
 packages = find:
