### 2024-01-29 calebweinreb/SNUB#13 @calebweinreb
SNUB is now available via pip https://pypi.org/project/systems-neuro-browser/

The name ``snub`` was already taken on PyPI, so to install, users will have to run ``pip install systems-neuro-browser`` although they will still ``import snub`` in python and run ``snub`` from the command line.

I chose not to implement automatic updates (e.g. via GitHub Actions) since I will remain the main developer for the foreseeable future and can perform the updates manually.
