### 2024-01-29 calebweinreb/SNUB#18 @niksirbi
[Major] Qt5 installation instructions don't work on new Ubuntu versions
The installation instructions for Linux include `sudo apt-get install qt5-default`, but the `qt-default` package is [no longer available for recent versions of Ubuntu](https://askubuntu.com/questions/1335184/qt5-default-not-in-ubuntu-21-04).

Running `sudo apt-get install qtbase5-dev qtchooser qt5-qmake qtbase5-dev-tools`, as responses to the above-linked thread suggest, didn't do the trick for me, and neither did `conda install pyqt`, `conda install qt` or `conda install qtpy`.

No matter what I do (on ubuntu 23.04 or 23.10) I end up with one of these two errors when I try to run `snub`:

```bash
QObject::moveToThread: Current thread (0x1e155c0) is not the object's thread (0x2da9de0).
Cannot move to target thread (0x1e155c0)

qt.qpa.plugin: Could not load the Qt platform plugin "xcb" in "/home/niko/anaconda3/envs/snub2/lib/python3.8/site-packages/cv2/qt/plugins" even though it was found.
This application failed to start because no Qt platform plugin could be initialized. Reinstalling the application may fix this problem.

Available platform plugins are: xcb, eglfs, minimal, minimalegl, offscreen, vnc, webgl.

Aborted (core dumped)
```

```bash
ImportError: $HOME/anaconda3/envs/snub3/bin/../lib/libstdc++.so.6: version `GLIBCXX_3.4.30' not found (required by /usr/lib/x86_64-linux-gnu/libicuuc.so.72)
```

There should be a less brittle way of installing `snub` on recent Linux distros.
