### 2024-01-29 calebweinreb/SNUB#13 @niksirbi
I can confirm that `pip install systems-neuro-browser` works! Thanks.
