### 2024-01-29 calebweinreb/SNUB#19 @niksirbi
[SUGGESTION] Consider making development dependencies optional
Some of the dependencies, e.g. `pytest` and `pytest-qt` are only required during development, and not needed for users. You could [make those dependencies optional](https://setuptools.pypa.io/en/latest/userguide/dependency_management.html#optional-dependencies) under an arbitrary identifier (e.g. `dev`) in your `setup.cfg` file.

```
[options.extras_require]
dev =
    pytest
    pytest-qt
```

This way `pip install systems-neuro-browser` will install only the smaller set of "normal" dependencies, whereas `pip install systems-neuro-browser[dev]` will also include the testing dependencies. In future, if you add linters and or type checking, you could also group those tools under the. `dev` optional dependencies.
