### 2024-01-29 calebweinreb/SNUB#20 @niksirbi
[MAJOR] Missing contributing guide
@calebweinreb you seem to be the only developer at the moment, but I assume that you would like for the project to grow and to inspire contributions from collaborators, and perhaps also from the wider systems neuro community.

I suggest adding a `CONTRIBUTING.md` guide to the repo and, at the minimum, stating your policy on external contributions. Are they welcome, and if so, what's the recommended procedure?
