### 2024-01-29 calebweinreb/SNUB#8 @niksirbi
I ran the tests locally and measured a test coverage of 54% (using `pytest-cov`). That's a great start!
I opened a separate issue #19 where I suggest separating the "dev" dependencies (e.g. `pytest`), from the ones users require.

Ideally, you would want to run these tests across OS platforms and Python versions (e.g. using GitHub's CI features), but I'm fine with you adding those later.
