### 2024-01-30 847327a
pip opencv version for ubuntu compatibility
