+++ setup.cfg
@@ -31,7 +31,7 @@ install_requires =
     ipykernel
     pyqtgraph
     networkx
-    opencv-python-headless
+    opencv-python-headless==4.1.2.30
     vidio>=0.0.3
     pytest
     pytest-qt
