### 2024-01-30 #6187 @emdupre
👋  Hi @calebweinreb, thanks for checking in on this !

> I believe I have now responded to all the github issues raised by the reviewers. There are a number of checklist items where one or more reviewer left the item unchecked but did not raise a corresponding issue. Is there a process for seeking additional feedback on these points and/or offering a rebuttal? Thanks!

Apologies if this was unclear : we had extended a month review window for this work, so we're still within that initial review period. It's great, though, that you're addressing issues as they're raised ! This helps to make the whole process faster, and I certainly appreciate seeing the progress there.

@niksirbi @vigji and @lucasmiranda42, thank you again for your input to date on _SNUB_ ! Please do let us know once you've finished your initial review, so we can make sure that your comments are fully considered.

And of course, please don't hesitate to ping in this issue if you have any other questions !
