### 2024-01-31 calebweinreb/SNUB#21 @niksirbi
[Major] Clarify provenance of sample datasets
Thanks for publishing the data on Zenodo @calebweinreb #10.

I couldn't find any information about the provenance of the sample datasets though. If these are taken from public data repositories, I suggest citing them. If these are original data, you would have to mention something about the licensing of animal experiments and confirm that it complies with the [JOSS guidelines on the subject](https://joss.readthedocs.io/en/latest/policies.html#joss-policies).

I would include the provenance info in the Zenodo data description, and cite the Zenodo dataset in the paper.
