### 2024-01-31 calebweinreb/SNUB#12 @niksirbi
> The primary use case we have in mind is that a person has collected data and begun to analyze it in jupyter. They are totally comfortable loading the data and manipulating it with python. They likely are already generating static versions of SNUB-like data views, such as aligned heatmaps and trace plots, and now would like a frictionless way to pan/zoom around these plots and keep all the data views linked together. So instead of a big matplotlib figure, they create a SNUB project. Just as the matplotlib figure is agnostic to any loading/preprocessing that has already happened, so are the SNUB i/o routines.

This clarifies it, at least for me. Perhaps you could put more emphasis on this point in the paper and in the GitHub README/docs.

Regarding data duplication during project creation, at least you provide the option to avoid that for videos (via the `copy` parameter). But I wonder whether it's worth having that set to `False` by default, since videos may be quite large. I would also consider providing the `copy=False` option for more data types, where possible, especially those that are likely to be large files.
