### 2024-01-31 calebweinreb/SNUB#11 @niksirbi
There is a related issue to that one.

As a user, I would like to be able to change visualisation parameters like `vmin`, `vmax`, `colormap`, `linewidth` etc. directly in the GUI. It's a bit weird to pre-specify these when adding data to the project, because at that point you haven't seen the data and you might not know what visualisation parameters are appropriate. Some of these parameters are editable from the GUI, but not all of them.
