### 2024-01-31 calebweinreb/SNUB#22 @niksirbi
[SUGGESTION] Ability to add and remove dataviews in the GUI
Currently, users have to first create a project by adding all relevant data sources via functions like `snub.io.add_video`, `snub.io.add_scatter` etc. Once a project is created and opened, there is no way to add more data sources as views directly in the GUI. It would be nice to initialise a project with a few views, and be able to add more as you see fit.

In other words, I would like for the `add_` methods to be exposed in the GUI, perhaps under the "File" menu.  It would also be nice to have the option to close data views in the GUI (e.g. via an 'X' button on each panel).

This is more of an idea for a possible future direction. Feel free to dismiss if you think it would complicate things too much.
