### 2024-01-31 calebweinreb/SNUB#23 @niksirbi
[Minor] Some small UI bugs and suggestions
When using the SNUB GUI on my M2 Macbook I encountered several small "glitches". I chose to group them all here instead of opening a separate issue for each. Some of these are "bugs", and other others are suggestions for UI/UX improvement. None of these are critical, but cumulatively they may lead to user frustration.

## Bugs

- Pressing the `Esc` button removes the UMAP scatterplot, without an obvious way to bring it back.
- In the calcium imaging sample project, selecting the row labels on the heatmap generates the corresponding trace, while this doesn't work for spike plots in the ephys project. "Right click -> Plot trace" works for both of them. As a user, I would expect identical behaviour from these two types of plots - i.e. clicking a row label should plot the trace in both cases.
- The play button and the "Track playhead" checkbox do not display properly for me.

![Screenshot 2024-01-30 at 18 10 06](https://github.com/calebweinreb/SNUB/assets/20923448/53ae1b1e-344d-43eb-a1c6-7ab771fda180)

## Suggestions

- Add a "Help" button or similar to show the most important shortcuts and functionalities. There are many useful tricks for using the interface (documented [here](https://snub.readthedocs.io/en/latest/tutorials.html#using-the-interface)), but they are easily forgotten till you get very familiar with them. A list of shortcuts. easily accessible from the GUI would facilitate the learning process.
- It is hard to select a specific cell in heatmaps and trace plots when there are many cells. In heatmaps the row labels are very closely bunched together, while in trace plots the dropdown gives you an enormous vertical list of cells. I don't know what the solution is, just marking it as a point that would benefit from improvement.
