### 2024-01-31 calebweinreb/SNUB#24 @niksirbi
[Minor] Some typos I found in the paper and the docs
## Paper
- Line 24: `dislay`
- Line 31: `snub.io.add_splikeplot`

## Docs
- Homepage: `snub.io.add_splikeplot`
- Calcium imaging tutorial: `There are also a behavior annotations` (the "a" should be removed).
