### 2024-01-31 #6187 @niksirbi
Overall I think SNUB is a great tool. Something like it is needed in systems neuroscience, as more and more researchers acquire multiple synchronized data streams from the same animal.

SNUB's primary purpose - as stated - is data exploration and visualisation and it achieves this goal very well. I was able to install SNUB and reproduce the tutorials using the provided sample data. I found the GUI to be performant and responsive, at least on my M2 Macbook.

I think SNUB has the potential for wider adoption especially if the installation issues are addressed and if the data loading is made smoother (see individual issues below).

In the following list I've included both issues raised by me and issues raised by other reviewers that I agree with. I consider the ones marked as `[Major]` to be blockers for the JOSS reviewer checklist. The others are not critical, and some are purely suggestions/ideas for the author's (optional) consideration.

@calebweinreb has already made inroads on addressing the major issues, so I'm confident that they will be solved. I'll happily recommend SNUB for publication in JOSS once the major issues are closed.

## Installation
- https://github.com/calebweinreb/SNUB/issues/6
- https://github.com/calebweinreb/SNUB/issues/18
- https://github.com/calebweinreb/SNUB/issues/13
- https://github.com/calebweinreb/SNUB/issues/19

Hiccups during installation can discourage new users and it's a shame if they miss out on SNUB because of them. The major issues should be resolved, or at least, the supported OS platforms (and versions thereof) should be indicated. @calebweinreb has already started working on making installation easier, e.g. by publishing the package on PyPI.

## Data loading
- https://github.com/calebweinreb/SNUB/issues/9
- https://github.com/calebweinreb/SNUB/issues/11
- https://github.com/calebweinreb/SNUB/issues/12
- https://github.com/calebweinreb/SNUB/issues/22

I think that making data loading as seamless as possible for the users will be critical for the project's long-term success. @calebweinreb has already made a big improvement by adding support for converting NWB files, as @vigji suggested.

## Maintainability and community
- https://github.com/calebweinreb/SNUB/issues/8
- https://github.com/calebweinreb/SNUB/issues/15
- https://github.com/calebweinreb/SNUB/issues/20

@calebweinreb seems to be the only active developer on the project. It would be useful to clarify to what extent contributions from others are welcome and what is the recommended process. If others start contributing to SNUB, issues like testing and linting will become increasingly important. Some tests have been already introduced, which is great 🎉 !

Although I haven't opened an issue about Continuous Integration, I would highly recommend adopting some sort of CI framework (e.g. GitHub actions) for automating tests on multiple OS platforms, which would also help catch the aforementioned installation problems.

## Miscellaneous
- https://github.com/calebweinreb/SNUB/issues/10
- https://github.com/calebweinreb/SNUB/issues/21
- https://github.com/calebweinreb/SNUB/issues/23
- https://github.com/calebweinreb/SNUB/issues/24

Thanks @emdupre for tagging me as a reviewer! It's a pleasure to review for JOSS.
