### 2024-02-01 calebweinreb/SNUB#22 @calebweinreb
I like this idea! Adding an "X" button for closing a widget is easy and will be included in the next release (implemented here https://github.com/calebweinreb/SNUB/pull/25).

Adding new widgets through the GUI is a little more of a lift so I may wait to see if it is requested by anyone else. But in principle I could definitely imagine a generic widget that programmatically exposes the args/kwargs for each respective add_ function.
