### 2024-02-01 6700bed
Prevent scatter from disappearing when ESC is pressed
