### 2024-02-01 b3c5fc2
connect spikeplot.heatmap_labels.display_trace_signal to "show_trace" slot
