### 2024-02-01 9306206
add snub/gui/icons/* to package data
