+++ setup.cfg
@@ -9,11 +9,13 @@ classifiers =
     Programming Language :: Python :: 3
     Operating System :: OS Independent
 
-
-
 [options]
 packages = find:
 include_package_data = True
+
+[options.package_data]
+snub.gui = icons/*
+
 python_requires = >=3.8
 install_requires =
     PyQt5
