### 2024-02-01 9e270e7
remove duplicate options.package_data in setup.cfg
