+++ setup.cfg
@@ -15,6 +15,7 @@ include_package_data = True
 
 [options.package_data]
 snub.gui = icons/*
+* = *.md
 
 python_requires = >=3.8
 install_requires =
@@ -48,9 +49,6 @@ install_requires =
 [options.entry_points]
 console_scripts =
     snub = snub.gui.main:run
-
-[options.package_data]
-* = *.md
 
 [versioneer]
 VCS = git
