+++ setup.cfg
@@ -12,11 +12,6 @@ classifiers =
 [options]
 packages = find:
 include_package_data = True
-
-[options.package_data]
-snub.gui = icons/*
-* = *.md
-
 python_requires = >=3.8
 install_requires =
     PyQt5
@@ -45,6 +40,9 @@ install_requires =
     pandas
     dandi
 
+[options.package_data]
+snub.gui = icons/*
+* = *.md
 
 [options.entry_points]
 console_scripts =
