### 2024-02-01 7ba804d
Fix second opencv-related bug in Ubuntu
