+++ setup.cfg
@@ -29,7 +29,7 @@ install_requires =
     ipykernel
     pyqtgraph
     networkx
-    opencv-python-headless==4.1.2.30
+    opencv-python-headless==4.5.5.64
     vidio>=0.0.3
     pytest
     pytest-qt
