### 2024-02-01 calebweinreb/SNUB#18 @calebweinreb
This turned out to be an opencv-related issue. I got ubuntu installation to work by switching the opencv requirement to `opencv-python==4.1.2.30`. Nothing needs to be installed via apt-get.

pip install should work now as of version 0.0.13
