### 2024-02-05 54f2599
allow adjustment of markersize in linewidth in scatter plot
