### 2024-02-05 180b88e
adjust y-axis and linewidth in traceplot + changed appearance of checkmark in context menu
