+++ snub/gui/utils/__init__.py
@@ -1,2 +1,8 @@
 from .interval import IntervalIndex
-from .widgets import AdjustColormapDialog, HeaderMixin, CheckBox
+from .widgets import (
+    AdjustColormapDialog,
+    HeaderMixin,
+    CheckBox,
+    CHECKED_ICON_PATH,
+    UNCHECKED_ICON_PATH,
+)
