### 2024-02-05 calebweinreb/SNUB#11 @calebweinreb
The overall rationale for ingesting data and setting plotting parameters in a single function is explained here https://github.com/calebweinreb/SNUB/issues/11

It is a good point that though that users might wish to make further adjustments once the SNUB plot is already open, especially ones that could potentially impact the legibility of the plot. I just merged https://github.com/calebweinreb/SNUB/pull/28 that makes a few more parameters adjustable in the GUI:
- scatter plot: users can now adjust marker size and border width directly in the GUI
- trace plot: users can adjust linewidth and y-axis bounds in the GUI

These parameter changes are not saved to the config (currently the GUI never directly changes the config and I'd like to keep it that way).
