### 2024-02-05 calebweinreb/SNUB#18 @niksirbi
Thanks @calebweinreb, I can confirm this works now! Good job getting rid of the apt-get dependency. I'm closing the issue.
