+++ paper.md
@@ -35,14 +35,14 @@ Direct inspection of behavior and neurophysiology recordings is hard because the
 
 ![Screenshot from SNUB.\label{fig:screenshot}](docs/media/screenshot.png)
 
-We provide dedicated widgets and loading functions for exploring raw video, 3D animal pose, behavior annotations, electrophysiology recordings, and calcium imaging data - either as a raster or as a super-position of labeled regions of interest (ROIs). More broadly, SNUB can dislay any data that takes the form of a heatmap, scatter plot, video, or collection of named temporally-varying signals. 
+We provide dedicated widgets and loading functions for exploring raw video, 3D animal pose, behavior annotations, electrophysiology recordings, and calcium imaging data - either as a raster or as a super-position of labeled regions of interest (ROIs). More broadly, SNUB can display any data that takes the form of a heatmap, scatter plot, video, or collection of named temporally-varying signals. 
 
 In addition to the front-end GUI, we include a library of functions that ingest raw data and save it to a format that is readable by the SNUB viewer. The following code, for example, creates a project with paired electrophysiology and video data.
 
 ```
 snub.io.create_project(project_directory, duration=1800)
 snub.io.add_video(project_directory, 'path/to/my_video.avi', name='IR_camera')
-snub.io.add_splikeplot(project_directory, 'my_ephys_data', spike_times, spike_labels)
+snub.io.add_spikeplot(project_directory, 'my_ephys_data', spike_times, spike_labels)
 ```
 
 We also provide a rudimentary tool for automatically generating SNUB datasets from Neurodata Without Borders (NWB) files, which contain raw and processed data from neuroscience recordings [@NWB]. The data in NWB files are stored hierarchically, and each component of the hierarchy has a specific neurodata type that reflects the measurement modality (e.g, "Units" for spike trains, "ImageSeries" for video). Our conversion tool generates a SNUB display element for each supported neurodata type. Users can optionally restrict this process to a subset of the NWB hierarchy (e.g., include pose tracking while excluding electrophysiology, or include just a subset of electrophysiology measurements). 
