+++ docs/source/tutorials.rst
@@ -10,7 +10,7 @@ Download the `example behavior and calcium imaging data <https://zenodo.org/reco
 project generated in this tutorial is available `here <https://zenodo.org/records/10578025/files/miniscope_project.zip?download=1>`_. 
 This tutorial is based on head-mounted 1-photon calcium imaging and video recordings
 from a mouse engaged in social interaction. The camera and microscope were synchronized 
-and have associated timestamps in seconds. There are also a behavior annotations
+and have associated timestamps in seconds. There are also behavior annotations
 for each frame generated by MoSeq. 
 
 
