### 2024-02-05 calebweinreb/SNUB#24 @calebweinreb
Wow thank you for catching these! Fixed here https://github.com/calebweinreb/SNUB/commit/389dde73608ff707c720cb2e053fc66d789083d6
