### 2024-02-05 7c13074
make pytest deps optional
