+++ setup.cfg
@@ -31,14 +31,17 @@ install_requires =
     networkx
     opencv-python==4.1.2.30
     vidio>=0.0.3
-    pytest
-    pytest-qt
     pynwb
     ndx-pose
     ndx-photometry
     ndx-depth-moseq
     pandas
     dandi
+
+[options.extras_require]
+dev = 
+    pytest
+    pytest-qt
 
 [options.package_data]
 snub.gui = icons/*
