### 2024-02-05 0b33f1a
mention dev install option in docs
