### 2024-02-05 cbb8030
mention dev install option in README
