### 2024-02-05 calebweinreb/SNUB#19 @calebweinreb
I implemented the dev option exactly as described (https://github.com/calebweinreb/SNUB/commit/7c130749962e90ddbc63207441aaa6dc203d5fa9). Its now available on PyPI for version >= 0.0.14
