### 2024-02-05 calebweinreb/SNUB#6 @vigji
I tested it over a different M1 machine and a Windows machine and everything was fine there. I do not know what funny issue causes this segmentation fault just on my machine, but I'll close this for now!
