### 2024-02-05 calebweinreb/SNUB#8 @vigji
I think this is a great start for testing, enough to at least check out basic functioning is someone wants to contribute to the project. Agreed that multiple platforms can come later, I'll close this for now!
