### 2024-02-05 calebweinreb/SNUB#9 @vigji
I see the point that you make in referring to it as an almost matplotlib-like tool in being data-agnostic. Still, I feel that including pipelines to generate the data from nwb files does lower the energy barrier for adoption for new users, and gives a good tool to dig into open datasets. I feel that this is addressed now!
