### 2024-02-05 calebweinreb/SNUB#12 @vigji
I understand the point of avoiding repeating every time pre-processing steps; this cost would remain in the no-copy configuration. Maybe in practical scenarios most users would just duplicate. Still, I agree on giving the option for not copying the data; again, not just for the sake of space but also to avoid potentially pernicious duplications that hide changes in the source data (although maybe for most uses source data would not change enough to justify this concern).
