+++ setup.cfg
@@ -42,6 +42,8 @@ install_requires =
 dev = 
     pytest
     pytest-qt
+    sphinx==4.4.0
+    sphinx_rtd_theme==1.0.0
 
 [options.package_data]
 snub.gui = icons/*
