### 2024-02-16 calebweinreb/SNUB#22 @vigji
I agree that this would be a great functionality to add!
