### 2024-02-16 #6187 @vigji
Hi @emdupre !

I have updated the status of my checklist based on the current state of the package.

Overall, I agree with @niksirbi on SNUB being a useful resource (it was actually already in use in my lab before the submission, so I am sure about the potential!). I am not aware of other tools that it compares to, and I believe that it fills a gap in providing seamless visualizations for multi-modality time series data.

I think that the scope and statement of need for the package are described with enough details in the paper. It is missing a session on comparison with other tools, but I am not aware of any so I do not have issues with that.

I do think that @calebweinreb implemented most of the changes that I was considering major for publication. There is now a version of testing that can be used for a cursory check on functionality by contributors, and nice use cases for DANDI datasets that will definitively highlight how flexibly the package could be used to explore the user' own data or even existing open access datasets.

There are still some minor changes and further improvements that can be done on the data input ( #12) and interface's side (#22), but I do not consider those blocking issues.

Blocking issues:
- I agree with @niksirbi on the need for a clearer statement for contributors (#20). To which degree are contributions welcome? Is there something like a roadmap for future development? I think there should be a clearer specification for this as per JOSS checklist.
- Finally, also per JOSS guidelines the provenance of the example data should be addressed (#21).

Once those are addressed, this is good for publication on my side!
