### 2024-02-16 #6187 @emdupre
Thank you, @vigji !

I assume you meant to link to https://github.com/calebweinreb/SNUB/issues/20 and https://github.com/calebweinreb/SNUB/issues/21, rather than the listed JOSS issues 😸

I will continue to track those and am looking forward to updates. Thank you again both for this summary and for your expertise in reviewing  _SNUB_ !
