+++ snub/gui/help/trace_plots.md
@@ -1,0 +1,19 @@
+## Trace plots
+
+Trace plots display time-series data using lines. They are aligned to the [timeline](timeline.md) at the bottom of the window and can be navigated using gestures (pan and zoom). 
+
+### Showing and hiding traces
+
+For each currently visible trace, a small color-matched label-box is displayed in the top-right.
+
+* **To show a trace**, click the drop-down menu in the top right corner and toggle the corresponding checkbox.
+* **To hide a trace**, either toggle its checkbox in the drop-down menu or click on the corresponding label-box in the top-right.
+* **To hide all traces**, right click on the trace plot and select "Hide all traces".
+
+### Adjusting the y-axis
+
+By default, the y-axis limits are determined automatically based on the currently visible traces; if traces are shown or hidden, the y-axis limits will be updated accordingly.
+
+* **To toggle automatic updating of y-axis limits**, right click on the trace plot and check or uncheck "Automatic y-axis limits". 
+* **To adjust y-axis limits**, right click on the trace plot and select "Adjust y-axis limits". This will temporarily disable automatic updating of y-axis limits. 
+
