+++ snub/gui/help/loading_data.md
@@ -1,0 +1,7 @@
+## Open a SNUB project
+
+To open a SNUB project, go to File > Open Project, navigate to the project directory, and hit "Choose" with the directory selected. Multiple projects can be opened at once as different tabs. Projects can also be opened by including their paths as command line argument(s) when launching SNUB.
+
+## Reload a project
+
+If a project is already open and its files have been modified outside of SNUB, the project can be reloaded in the SNUB GUI by going to File > Reload Project.
