+++ setup.cfg
@@ -37,6 +37,8 @@ install_requires =
     ndx-depth-moseq
     pandas
     dandi
+    markdown
+    PyQtWebEngine
 
 [options.extras_require]
 dev = 
