### 2024-02-22 bf7127a
- Set copy=False as default for videos - Allow passing heatmap as a path - Allow passing spike data as a path
