+++ snub/io/project.py
@@ -5,6 +5,7 @@ import shutil
 import warnings
 import pickle
 import colorsys
+import cmapy
 import scipy.sparse
 from vidio import VideoReader
 
@@ -342,9 +343,14 @@ def edit_dataview_properties(project_directory, dataview_type, name, **kwargs):
     save_config(project_directory, config)
 
 
+def _samepath(path1, path2):
+    normalized_path1 = os.path.normcase(os.path.realpath(path1))
+    normalized_path2 = os.path.normcase(os.path.realpath(path2))
+    return normalized_path1 == normalized_path2
+
+
 def remove_dataview(project_directory, dataview_type, name, delete_data=False):
-    """Remove a data-view from the specified project and (optionally)
-    delete its data.
+    """Remove a data-view from the specified project and (optionally) delete its data.
 
     Parameters
     ----------
@@ -360,8 +366,9 @@ def remove_dataview(project_directory, dataview_type, name, delete_data=False):
     delete_data: bool, default=False
         Delete the data associated with the given data-view, which may
         be .avi, .npy or .hdf5 files depending on the type of data-view.
+        This is automatically prevented if the data is located outside
+        the project directory.
     """
-
     config = load_config(project_directory)
     index = _get_named_dataview_index(config, dataview_type, name)
     if index is None:
@@ -373,10 +380,19 @@ def remove_dataview(project_directory, dataview_type, name, delete_data=False):
     if delete_data:
         for key, value in config[dataview_type][index].items():
             if "path" in key:
-                data_path = os.path.join(project_directory, value)
-                if os.path.exists(data_path):
-                    print("Deleting", data_path)
-                    os.remove(data_path)
+                if os.path.dirname(value) == "":
+                    # relative path in the project directory
+                    print("Deleting", value)
+                    os.remove(os.path.join(project_directory, value))
+                elif _samepath(os.path.dirname(value), project_directory):
+                    # absolute path in the project directory
+                    print("Deleting", value)
+                    os.remove(value)
+                else:
+                    print(
+                        f"The data for this {dataview_type} is located outside the "
+                        "project directory and will not be deleted."
+                    )
     del config[dataview_type][index]
     print('Removed {} with the name "{}"'.format(dataview_type, name))
     save_config(project_directory, config)
@@ -385,7 +401,7 @@ def remove_dataview(project_directory, dataview_type, name, delete_data=False):
 def add_video(
     project_directory,
     videopath,
-    copy=True,
+    copy=False,
     name=None,
     fps=30,
     start_time=0,
@@ -402,48 +418,43 @@ def add_video(
         Project that the video should be added to.
 
     videopath : str
-        Path to an existing video. The video will be read as 8bit RGB.
-        Other video formats, such as 16bit depth or 16bit monochrome,
-        will have to be converted to 8bit RGB before they can be added.
-        Functions for doing so are provided in :py:mod:`snub.io.video_conversion`.
+        Path to an existing video. The video will be read as 8bit RGB. Other video
+        formats, such as 16bit depth or 16bit monochrome, will have to be converted to
+        8bit RGB before they can be added. Functions for doing so are provided in
+        :py:mod:`snub.io.video_conversion`.
 
-    copy: bool, default=True
-        It is recommended that all data for a given project is contained
-        within the project directory. Leave ``copy=True`` if the video file
-        is currently outside the project directory and you wish to copy it.
-        Otherwise if ``copy=False``, the config will store the relative path
-        from the project directory to the video file.
+    copy: bool, default=False
[…]
