### 2024-02-22 7287ca3
clarfied that SNUB is like a plotting library in the JOSS manuscript
