### 2024-02-22 calebweinreb/SNUB#12 @calebweinreb
These are good points! To avoid unnecessary data duplication, I've set `copy=False` as the default for videos, and now give users the option to provide heatmap or spiking data as a path (which is not copied) rather than an array (see https://github.com/calebweinreb/SNUB/commit/bf7127aeed030c0fffbdccabf54ca41955625df2). This option is still not available for the other GUI elements (scatter plots, trace plots, etc.) because that would have required more substantive reorganization of how the data for those elements is stored and read by the GUI. Still, I think this addresses most of the space issue since heatmaps and spiking data are usually the heftiest files by an order of magnitude.

I also [updated the JOSS manuscript](https://github.com/calebweinreb/SNUB/commit/7287ca3a3f9662b7c5f36b8ce024e27715499fa1) by adding the text below to clarify that SNUB is akin to a plotting library, as noted in the discussion above.

*The main intended user is a researcher who has collected data and started to analyze it (e.g. in python). They may already be generating static versions of the visualizations afforded by SNUB, such as aligned heatmaps and trace plots, and now would like a frictionless way to pan/zoom around these plots and keep all the data views linked together. Importantly, SNUB should be thought of as akin to a plotting library, rather than a data analysis tool.*
