### 2024-02-22 calebweinreb/SNUB#23 @calebweinreb
Thanks for the detailed list! I believe I'd addressed all the points above (except perhaps suggestion bullet 2).

- Pressing `Esc` no longer hides the scatter plot (https://github.com/calebweinreb/SNUB/pull/26)
- Spike plots now behave like heatmaps: clicking on row labels causes the corresponding trace to appear (https://github.com/calebweinreb/SNUB/pull/27/commits/b3c5fc25a3691fdd6d5016bf28eea5170ecc7e20)
- This absence of a play button and other icons was related to the corresponding icon images not being included in the package data. This is now fixed (https://github.com/calebweinreb/SNUB/pull/27/commits/93062068a97af4502ededab25ea8d050aa3dae80).
- I added a help menu that includes comprehensive explanations of all the tips/tricks for navigating the GUI (see  https://github.com/calebweinreb/SNUB/pull/29 and screenshots).
- The bunching of labels in the heatmap is a good point. I think vertical zoom and other features make it pretty easy to go from a row of the heatmap to a row label, but it's quite hard to go in the other direction, finding a row from its label. However, I think this wouldn't be such a common use case e.g. when viewing large-population neural recordings.
- I think the giant drop-down menu for the trace plot is more manageable since the labels are sorted alphabetically.
