### 2024-02-22 calebweinreb/SNUB#23 @niksirbi
Thanks for implementing all the suggestions! The help menu especially looks great!

I tried to double-check that these work on my machine by pip installing from the current `main` branch (in a fresh conda environment with Python 3.8). But the dependencies failed to resolve. Could you double check that some of the newly introduced dependencies do not cause version incompatibilities?
