+++ snub/gui/help/__init__.py
@@ -1,0 +1,1 @@
+from .help_menu import HelpMenu
