### 2024-02-23 calebweinreb/SNUB#23 @calebweinreb
Thanks for checking! I tried installing from pip (which contained the equivalent of the main branch) with no issues on Mac and Linux (incidentally there was a circular import error when I tried to run SNUB, but that us now fixed).

What OS were you using? And what error was produced?
