+++ docs/source/conf.py
@@ -12,14 +12,15 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../../'))
+
+sys.path.insert(0, os.path.abspath("../../"))
 
 
 # -- Project information -----------------------------------------------------
 
-project = 'snub'
-copyright = '2022, Caleb Weinreb'
-author = 'Caleb Weinreb'
+project = "snub"
+copyright = "2022, Caleb Weinreb"
+author = "Caleb Weinreb"
 
 
 # -- General configuration ---------------------------------------------------
@@ -28,16 +29,17 @@ author = 'Caleb Weinreb'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-	'sphinx.ext.autodoc',
-	'sphinx.ext.napoleon',
-	'sphinx.ext.autosummary',
-	'sphinx.ext.viewcode',
-	'sphinx.ext.coverage',
-	'sphinx.ext.autosectionlabel']
+    "sphinx.ext.autodoc",
+    "sphinx.ext.napoleon",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.viewcode",
+    "sphinx.ext.coverage",
+    "sphinx.ext.autosectionlabel",
+]
 
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -50,16 +52,15 @@ exclude_patterns = []
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
+html_theme = "sphinx_rtd_theme"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ["_static"]
 
-html_favicon = '../media/favicon.png'
+html_favicon = "../media/favicon.png"
+
 
 def setup(app):
-    app.add_css_file('css/custom.css')  # may also be an URL
-
-
+    app.add_css_file("css/custom.css")  # may also be an URL
