+++ setup.cfg
@@ -43,6 +43,7 @@ install_requires =
 [options.extras_require]
 dev = 
     pytest
+    black
     pytest-qt
     sphinx==4.4.0
     sphinx_rtd_theme==1.0.0
