### 2024-02-23 calebweinreb/SNUB#20 @calebweinreb
I've now added a contribution guide:

https://github.com/calebweinreb/SNUB/blob/main/CONTRIBUTING.md
