+++ paper.bib
@@ -122,3 +122,16 @@
   issn         = {2050-084X},
   publisher    = {eLife Sciences Publications, Ltd}
 }
+
+@dataset{weinreb_2024_10578025,
+  author       = {Weinreb, Caleb and
+                  Osman, Mohammed Abdal Monium and
+                  Jay, Maya and
+                  Datta, Sandeep Robert},
+  title        = {Systems Neuro Browser (SNUB) example datasets},
+  month        = jan,
+  year         = 2024,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.10578025},
+  url          = {https://doi.org/10.5281/zenodo.10578025}
+}
