### 2024-02-23 calebweinreb/SNUB#21 @calebweinreb
I have updated the JOSS paper and the Zenodo description to clarify the source of the data used in the tutorial (see https://github.com/calebweinreb/SNUB/commit/4717b600c29a01c3549b1500812e5e81eb9d9bd7). The data are original and as yet unpublished. Here is a list of relevant changes:

- The author list of the JOSS paper and the Zenodo dataset have been updated to include the colleague who collected one of the sample datasets (I collected the other one).
- The JOSS paper notes that the data are original, includes an attestation of ethical compliance, and cites/links the Zenodo dataset
- The Zenodo data description also notes that the data are original, includes an ethical compliance statement, and also has a brief description of the experimental procedures for animal surgery, data collection and post-processing.
