### 2024-02-23 calebweinreb/SNUB#23 @niksirbi
Oh I see, I had tried to install from git by `pip install -e .` in the local repository's root, because I thought the PyPI version (0.0.3) is way behind the GitHub release.
I tried it again now, this time installing from pip and it worked! I can also verify that your UI fixes have worked, so we can close this issue.

I'm on Apple M2, Sonoma 14.3.1 btw. But I thinks this is not so relevant anymore. The documented way of installation clearly works.
