### 2024-02-26 #6187 @niksirbi
All of my issues have been addressed by @calebweinreb. I'll be happy to see this published in JOSS soon 🎉
