### 2024-02-29 calebweinreb/SNUB#30 @lucasmiranda42
[JOSS] Implement automated testing via CI
I think SNUB is a great tool, and has the potential to be used by quite some people!

That said, and especially given this is the case, I would strongly recommend implementing automated testing via CI (using something like [travis-ci](https://www.travis-ci.com/), for example). I think this will make keeping track of potential issues much easier as the package grows.

This issue is part of a paper review for [JOSS](https://github.com/openjournals/joss-reviews/issues/6187).
