### 2024-02-29 #6187 @lucasmiranda42
I successfully managed to install and reproduce all tutorials in macOS. I think SNUB is a great addition to the field, and has the potential to be used by quite some people!

That said, and especially given this is the case, I would strongly recommend implementing automated testing via CI (@calebweinreb, you could use set up something like [travis-ci](https://www.travis-ci.com/), for example).

Other than this, I just have a few minor comments on the paper (which in general reads very well!):

1) There seems to be a typo in line 24: it says "dislay" instead of "display"
2) I would suggest formatting the code between lines 29 and 31 as a python code box, which would read much nicer:

```python
snub.io.create_project(project_directory, duration=1800)
snub.io.add_video(project_directory, 'path/to/my_video.avi', name='IR_camera')
snub.io.add_splikeplot(project_directory, 'my_ephys_data', spike_times, spike_labels)
```

3) This may be a matter of taste, but the icon description at the very end reads a bit out of place. I would suggest moving it to the acknowledgements section.

Other than that, I think SNUB is ready to be published in JOSS!  🎉

Best,
Lucas
