### 2024-03-02 65d0d28
added a CI workflow for pytest and black formatting
