+++ .github/workflows/ci.yml
@@ -1,0 +1,45 @@
+name: CI
+
+on:
+    push:
+      branches:
+      - main
+  
+    pull_request:
+      branches:
+      - main
+
+    workflow_dispatch:
+
+jobs:
+  build-and-test:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-latest]
+        python-version: [3.9]
+
+    steps:
+    - uses: actions/checkout@v2
+
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v2
+      with:
+        python-version: ${{ matrix.python-version }}
+
+    - name: Install dependencies
+      run: |
+        pip install -e '.[dev]'
+
+    - name: Check Black Formatting
+      run: |
+        black --check .
+
+    - name: Run tests
+      run: |
+        # For GUI tests on Linux, you need to set up xvfb (X Virtual Framebuffer)
+        if [ "${{ runner.os }}" == "Linux" ]; then
+          sudo apt-get install -y xvfb libxkbcommon-x11-0
+          xvfb-run --auto-servernum --server-args="-screen 0 1024x768x24" pytest
+        else
+          pytest
