### 2024-03-02 ec572b3
only test on ubuntu, fix pyqt5 accommodations
