+++ .github/workflows/ci.yml
@@ -1,31 +1,33 @@
 name: CI
 
 on:
-    push:
-      branches:
-      - main
-  
-    pull_request:
-      branches:
-      - main
-
-    workflow_dispatch:
+  push:
+    branches:
+    - main
+  pull_request:
+    branches:
+    - main
+  workflow_dispatch:
 
 jobs:
   build-and-test:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest, windows-latest, macos-latest]
-        python-version: [3.8]
+    runs-on: ubuntu-latest
+    env:
+      DISPLAY: ':99.0'
 
     steps:
     - uses: actions/checkout@v2
 
-    - name: Set up Python ${{ matrix.python-version }}
+    - name: Set up Python
       uses: actions/setup-python@v2
       with:
-        python-version: ${{ matrix.python-version }}
+        python-version: 3.8
+
+    - name: Install system dependencies
+      run: |
+        sudo apt-get update
+        sudo apt-get install -y libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xfixes0 x11-utils
+        /sbin/start-stop-daemon --start --quiet --pidfile /tmp/custom_xvfb_99.pid --make-pidfile --background --exec /usr/bin/Xvfb -- :99 -screen 0 1920x1200x24 -ac +extension GLX
 
     - name: Install dependencies
       run: |
@@ -37,9 +39,4 @@ jobs:
 
     - name: Run tests
       run: |
-        # For GUI tests on Linux, you need to set up xvfb (X Virtual Framebuffer)
-        if [ "${{ runner.os }}" == "Linux" ]; then
-          sudo apt-get install -y xvfb libxkbcommon-x11-0
-          xvfb-run --auto-servernum --server-args="-screen 0 1024x768x24" pytest
-        else
-          pytest
+        pytest
