### 2024-03-02 840ad71
intentionally ill-formed setup.cfg to test github action
