+++ setup.cfg
@@ -12,7 +12,7 @@ classifiers =
 [options]
 packages = find:
 include_package_data = True
-python_requires = 3.8
+python_requires = >=3.8, <3.9
 install_requires =
     PyQt5
     numpy
