### 2024-03-02 calebweinreb/SNUB#30 @calebweinreb
Thanks for the suggestion! I created a github action that runs pytest and checks black formatting on pushes/PRs to `main`. I also created protections for `main` so that all merges require a PR.

Here's the github action config: https://github.com/calebweinreb/SNUB/blob/main/.github/workflows/ci.yml
Here's a log of the last completed action: https://github.com/calebweinreb/SNUB/actions/runs/8125132061/job/22207389372
