### 2024-03-03 1c973ad
format/spelling fixes suggested by lucasmiranda42 (https://github.com/openjournals/joss-reviews/issues/6187#issuecomment-1971339925)
