### 2024-03-05 #6187 @calebweinreb
Thanks @lucasmiranda42 for all your suggestions and the close reading! I'm not sure if this is the right place to respond, but I added automatic testing (described here https://github.com/calebweinreb/SNUB/issues/30) and implemented suggestions 1-3 above (see https://github.com/calebweinreb/SNUB/commit/1c973add316eeb470ca6482776b2e50ad24396e1 and current [paper draft](https://github.com/calebweinreb/SNUB/actions/runs/8132058349/artifacts/1293053042))
