### 2024-03-06 #6187 @emdupre
Thank you, @lucasmiranda42 for your review and @calebweinreb for your quick follow-up !

@lucasmiranda42, if you could confirm that these changes address your concerns, I'll be able to officially close the review 🚀

In the meantime, I'll perform a few editorial checks to make sure everything is on order for my checklist.
