### 2024-03-11 #6187 @lucasmiranda42
Dear @calebweinreb, dear @emdupre,

I can confirm that my concerns have been addressed. I have no objections to publication in JOSS :)

Best!
Lucas
