### 2024-03-11 #6187 @emdupre
Thank you, @lucasmiranda42 ! I noticed there are still a couple of unchecked boxes on your and @vigji's reviewer checklists. I'll plan to check these to show that the review is complete, given that you've both confirmed your reviews. If you have any concerns, please speak up !

Otherwise, I've also reviewed _Systems Neuro Browser_ @calebweinreb and am happy with the state of the submission. I do, though, have a few small editorial requests before we can proceed:

- [ ] In the listed author affiliations, please explicitly write out all acronyms; i.e., "MA" should be "Massachusetts." Please also include the country for each affiliation such that all affiliations follow "Department, Institute/Univ., City, Country" as closely as possible.
- [ ] As `rastermap` requests to cite the associated bioRxiv paper in their GitHub repository, please update to the included citation to the following:
   ```
   @article {Stringer2023.07.25.550571,
	author = {Carsen Stringer and Lin Zhong and Atika Syeda and Fengtong Du and Maria Kesa and Marius Pachitariu},
	title = {Rastermap: a discovery method for neural population recordings},
	elocation-id = {2023.07.25.550571},
	doi = {10.1101/2023.07.25.550571},
	publisher = {Cold Spring Harbor Laboratory},
	eprint = {https://www.biorxiv.org/content/early/2023/08/07/2023.07.25.550571.full.pdf},
	journal = {bioRxiv}
   }
   ```
- [ ] For the UMAP reference, it would be worth considering whether [the JOSS reference](https://joss.theoj.org/papers/10.21105/joss.00861) is the more appropriate citation (since you are using the tooling rather than discussing the mathematical underpinnings). If this is true, you could use the updated reference:
   ```
   @article{McInnes2018,
      doi = {10.21105/joss.00861},
      year = {2018}, publisher = {The Open Journal},
      number = {29}
      pages = {861},
      author = {Leland McInnes and John Healy and Nathaniel Saul and Lukas Großberger},
      title = {UMAP: Uniform Manifold Approximation and Projection},
      journal = {Journal of Open Source Software}
   }
   ```
- [ ] On line 15, as currently written there's a comma-splice at "high-dimensional, come in a variety of different modalities". Please either (1) replace the comma with an "and" or (2) revise line 16 from "with different sampling rates" to ", and have different sampling rates".
- [ ] Please capitalize Python throughout whenever referring to the programming language; e.g., line 19, line 68 in the VidIO reference
- [ ] On line 29, please replace the hypen with an em dash. Thus "data - either" --> "data—either"
- [ ] Please confirm the capitalization of VidIO on line 55
- [ ] Please confirm the citation formatting on line 61; i.e., why a single left square-bracket is rendering. I believe the raw text should be update from `[@weinreb_2024_10578025)` to `[@weinreb_2024_10578025]`. I would also remove the hyperlink on the text for Zenodo immediately preceeding this reference, as that URL is included in the `@weinreb_2024_10578025` reference.
