### 2024-03-13 5825463
update author affiliation per https://github.com/openjournals/joss-reviews/issues/6187#issuecomment-1989358973
