+++ paper.md
@@ -19,7 +19,7 @@ authors:
     orcid: 0000-0002-8068-3862
     affiliation: 1
 affiliations:
- - name: Department of Neurobiology, Harvard University, Boston, MA
+ - name: Department of Neurobiology, Harvard University, Boston, Massachusetts, United States of America
    index: 1
 date: 28 January 2024
 bibliography: paper.bib
