### 2024-03-13 35d3850
update JOSS paper referneces, per https://github.com/openjournals/joss-reviews/issues/6187#issuecomment-1989358973
