+++ paper.bib
@@ -62,14 +62,16 @@
   bdsk-url-1    = {https://doi.org/10.7554/eLife.63720}
 }
 
-
-@misc{rastermap,
-  author    = {C. Stringer and M. Pachitariu},
-  title     = {rastermap},
-  year      = {2020},
-  publisher = {GitHub},
-  journal   = {GitHub repository},
-  url       = {https://github.com/MouseLand/rastermap}
+@article {Stringer2023.07.25.550571,
+ author = {Carsen Stringer and Lin Zhong and Atika Syeda and Fengtong Du and Maria Kesa and Marius Pachitariu},
+ title = {Rastermap: a discovery method for neural population recordings},
+ elocation-id = {2023.07.25.550571},
+ year = {2023},
+ doi = {10.1101/2023.07.25.550571},
+ publisher = {Cold Spring Harbor Laboratory},
+ URL = {https://www.biorxiv.org/content/early/2023/08/07/2023.07.25.550571},
+ eprint = {https://www.biorxiv.org/content/early/2023/08/07/2023.07.25.550571.full.pdf},
+ journal = {bioRxiv}
 }
 
 @misc{vidio,
@@ -81,16 +83,17 @@
   url       = {https://github.com/jbohnslav/vidio}
 }
 
-@misc{umap,
-  doi       = {10.48550/ARXIV.1802.03426},
-  url       = {https://arxiv.org/abs/1802.03426},
-  author    = {McInnes, Leland and Healy, John and Melville, James},
-  keywords  = {Machine Learning (stat.ML), Computational Geometry (cs.CG), Machine Learning (cs.LG), FOS: Computer and information sciences, FOS: Computer and information sciences},
-  title     = {UMAP: Uniform Manifold Approximation and Projection for Dimension Reduction},
-  publisher = {arXiv},
-  year      = {2018},
-  copyright = {arXiv.org perpetual, non-exclusive license}
-}
+@article{umap, 
+   doi = {10.21105/joss.00861},
+   url = {https://doi.org/10.21105/joss.00861},
+   year = {2018}, publisher = {The Open Journal},
+   volume = {3},
+   number = {29}
+   pages = {861},
+   author = {Leland McInnes and John Healy and Nathaniel Saul and Lukas Großberger},
+   title = {UMAP: Uniform Manifold Approximation and Projection},
+   journal = {Journal of Open Source Software}
+} 
 
 @misc{petrucco_2020_3925903,
   author    = {Petrucco, Luigi},
