### 2024-03-13 ed41164
update references and typos per  https://github.com/openjournals/joss-reviews/issues/6187#issuecomment-1989358973
