+++ paper.bib
@@ -74,13 +74,12 @@
  journal = {bioRxiv}
 }
 
-@misc{vidio,
-  author    = {J. Bohnslav},
-  title     = {VidIO: simple, performant video reading and writing in python},
-  year      = {2020},
-  publisher = {GitHub},
-  journal   = {GitHub repository},
-  url       = {https://github.com/jbohnslav/vidio}
+@software{vidio,
+  author = {Bohnslav, Jim},
+  title = {VidIO: Simple, performant video reading and writing in python},
+  url = {https://github.com/jbohnslav/vidio},
+  version = {0.0.4},
+  date = {2024-03-11},
 }
 
 @article{umap, 
