+++ paper.bib
@@ -87,7 +87,7 @@
    url = {https://doi.org/10.21105/joss.00861},
    year = {2018}, publisher = {The Open Journal},
    volume = {3},
-   number = {29}
+   number = {29},
    pages = {861},
    author = {Leland McInnes and John Healy and Nathaniel Saul and Lukas Großberger},
    title = {UMAP: Uniform Manifold Approximation and Projection},
