### 2024-03-16 #6187 @calebweinreb
@emdupre thanks for the detailed edits! I made all the requested changes. The final version tag is [0.1.0](https://github.com/calebweinreb/SNUB/releases/tag/0.1.0)

The Zenodo DOI is 10.5281/zenodo.10825136
