### 2024-03-18 #6187 @emdupre
Thank you, @calebweinreb !

I have a two remaining requests on the archive itself :

- The version tag on the software repository is "SNUB 0.1.0" by on the archive it's listed as "Version v1". Could you please reconcile these ? This might be possible by [updating the archive metadata](https://help.zenodo.org/faq/technical/), or it might require minting a new release. If the latter is necessary, could you please provide me the new DOI ?
- It's currently using the default Zenodo license of CC-BY 4.0; however, _SNUB_ itself is licensed with an MIT license. Could you please update the license on the archive ? You should be able to do this directly by [updating the archive metadata](https://help.zenodo.org/faq/technical/).
