### 2024-03-22 #6187 @calebweinreb
I updated the version number and license and the DOI appears not to have changed.

(link for convenience: https://zenodo.org/records/10825136)
