### 2024-03-22 #6187 @emdupre
Thank you, @calebweinreb ! I'm now happy to recommend _SNUB_ to the EiC team for publication in JOSS ✨🚀

Thank you, too, to @vigji, @niksirbi, and @lucasmiranda42 for your reviews ! JOSS works because of your work 💐
