### 2024-03-24 #6187 @Kevin-Mattheus-Moerman
@calebweinreb as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Everything seems in order, there is only the following minor point that requires your attention:

- [ ] The affiliation says `Harvard University, Boston`, whereas I think this is usually `Harvard University, Cambridge`. We leave the affiliation up to you, just checking to make sure this is not an error.

@vigji and @lucasmiranda42, some of boxes remain unticked. If you feel these should be ticked please do so.

Relating to the unticked boxes, from inspecting this review and the repository/paper I can see that:
* The project has contributing guidelines
* Features sufficient testing
* The work complies with our Human and animal research policies (the paper reports on ethical approval for animal data/research).
* Language/State of the field. The paper appears in order.
The above was also confirmed with @emdupre via Slack, hence we will shortly proceed with acceptance of this submission, once the authors come back on the minor point regarding the affiliation.
