+++ paper.md
@@ -19,9 +19,9 @@ authors:
     orcid: 0000-0002-8068-3862
     affiliation: 1
 affiliations:
- - name: Department of Neurobiology, Harvard University, Boston, Massachusetts, United States of America
+ - name: Department of Neurobiology, Harvard Medical School, Boston, Massachusetts, United States of America
    index: 1
-date: 28 January 2024
+date: 24 March 2024
 bibliography: paper.bib
 
 ---
