### 2024-03-24 166a9dd
add gui.rst to docs
